+++ src/AbstractMaybe.php
@@ -20,64 +20,45 @@
   # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
   # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-  namespace Maybe;
+  namespace HaskellCamargo\Maybe;
 
-  # `Just` is a constructor of the `Maybe` type/monad. It must carry a value.
-  class Just extends Maybe implements IMaybe {
-    function __construct($value)
-    {
-      $this->value = $value;
-    }
-
+  # A `Maybe` type encapsulates an optional value. A value of type `Maybe a`
+  # either contains a value of type a (represented as `Just a`), or it is empty
+  # (represented as `Nothing`). Using `Maybe` is a good way to deal with errors
+  # or exceptional cases without resorting to drastic measures such as
+  # `Exception`.
+  # The `Maybe` type is also a monad. It is a simple kind of error monad, where
+  # all errors are represented by `Nothing`. A richer error monad can be built
+  # using the `Either` type.
+  abstract class AbstractMaybe {
     # Equivalent to Haskell's `>>=` operator. Its first argument is a value in
     # a monadic type, its second argument is a function that maps from the
     # underlying type of the first argument to another monadic type, and its
     # results is in that other monadic type.
-    function bind($fn) # :: (Maybe a, callable) -> Maybe b
-    {
-      return Maybe($fn($this->value));
-    }
+    abstract function bind($fn); # :: (Maybe a, callable) -> Maybe b
 
     # Extracts the element out of a `Just` and returns an error if its argument
     # is `Nothing`.
-    function fromJust() # :: Maybe a -> a
-    {
-      return $this->value;
-    }
+    abstract function fromJust(); # :: Maybe a -> a
 
     # Takes a `Maybe` value and a default value. If the `Maybe` is `Nothing`, it
     # returns the default values; otherwise, it returns the value contained in
     # the `Maybe`.
-    function fromMaybe($_) # :: (Maybe a, a) -> a
-    {
-      return $this->value;
-    }
+    abstract function fromMaybe($def); # :: (Maybe a, a) -> a
 
     # Returns true if its argument is of the form `Just _`.
-    function isJust() # :: Maybe a -> boolean
-    {
-      return true;
-    }
+    abstract function isJust(); # :: Maybe a -> boolean
 
-    # Returns true if its arguments is of the form `Nothing`.
-    function isNothing() # :: Maybe a -> boolean
-    {
-      return false;
-    }
+    # Returns false if its arguments is of the form `Nothing`.
+    abstract function isNothing(); # :: Maybe a -> boolean
 
     # Takes a default value, a function and, of course, a `Maybe` value. If the
     # `Maybe` value is `Nothing`, the function returns the default value.
     # Otherwise, it applies the function to the value inside the `Just` and
     # returns the result.
-    function maybe($_, $fn)  # :: (Maybe a, b, callable) -> b
-    {
-      return $fn($this->value);
-    }
+    abstract function maybe($def, $fn); # :: (Maybe a, b, callable) -> b
 
     # Returns an empty list when given ``Nothing`` or a singleton list when not
     # given ``Nothing``.
-    function toList() # :: Maybe a -> array
-    {
-      return [$this->value];
-    }
+    abstract function toList(); # :: Maybe a -> array
   }
