+++ src/IMaybe.php
@@ -20,7 +20,7 @@
   # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
   # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-  namespace Maybe;
+  namespace HaskellCamargo\Maybe;
 
   interface IMaybe {
     function bind($fn);                   # :: (Maybe a, callable) -> Maybe b
