+++ src/Just.php
@@ -20,54 +20,64 @@
   # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
   # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-  namespace Maybe;
-  use \Exception;
+  namespace HaskellCamargo\Maybe;
 
-  # `Nothing` is a constructor of the `Maybe` type/monad. It doesn't take value.
-  class Nothing extends Maybe implements IMaybe {
+  # `Just` is a constructor of the `Maybe` type/monad. It must carry a value.
+  class Just extends Maybe implements IMaybe {
+    function __construct($value)
+    {
+      $this->value = $value;
+    }
 
     # Equivalent to Haskell's `>>=` operator. Its first argument is a value in
     # a monadic type, its second argument is a function that maps from the
     # underlying type of the first argument to another monadic type, and its
     # results is in that other monadic type.
-    function bind($_) { # :: (Maybe a, callable) -> Maybe b
-      return $this;
+    function bind($fn) # :: (Maybe a, callable) -> Maybe b
+    {
+      return Maybe($fn($this->value));
     }
 
     # Extracts the element out of a `Just` and returns an error if its argument
     # is `Nothing`.
-    function fromJust() { # :: Maybe a -> a
-      throw new Exception("Cannot cal fromJust() on Nothing");
+    function fromJust() # :: Maybe a -> a
+    {
+      return $this->value;
     }
 
     # Takes a `Maybe` value and a default value. If the `Maybe` is `Nothing`, it
     # returns the default values; otherwise, it returns the value contained in
     # the `Maybe`.
-    function fromMaybe($def) { # :: (Maybe a, a) -> a
-      return $def;
+    function fromMaybe($_) # :: (Maybe a, a) -> a
+    {
+      return $this->value;
     }
 
-    # Returns false if its argument is of the form `Just _`.
-    function isJust() { # :: Maybe a -> boolean
-      return false;
+    # Returns true if its argument is of the form `Just _`.
+    function isJust() # :: Maybe a -> boolean
+    {
+      return true;
     }
 
     # Returns true if its arguments is of the form `Nothing`.
-    function isNothing() { # :: Maybe a -> boolean
-      return true;
+    function isNothing() # :: Maybe a -> boolean
+    {
+      return false;
     }
 
     # Takes a default value, a function and, of course, a `Maybe` value. If the
     # `Maybe` value is `Nothing`, the function returns the default value.
     # Otherwise, it applies the function to the value inside the `Just` and
     # returns the result.
-    function maybe($def, $_) { # :: (Maybe a, b, callable) -> b
-      return $def;
+    function maybe($_, $fn)  # :: (Maybe a, b, callable) -> b
+    {
+      return $fn($this->value);
     }
 
     # Returns an empty list when given ``Nothing`` or a singleton list when not
     # given ``Nothing``.
-    function toList() { # :: Maybe a -> array
-      return [];
+    function toList() # :: Maybe a -> array
+    {
+      return [$this->value];
     }
   }
